--- custom_components/ave_dominaplus/webserver/__init__.py
@@ -1,1 +1,0 @@
-"""Internal webserver package for AVE dominaplus."""
